move deps to types
